+++ dev-packages/application-manager/src/generator/backend-generator.ts
@@ -77,13 +77,13 @@ const main = require('@theia/core/lib/node/main');
 BackendApplicationConfigProvider.set(${this.prettyStringify(this.pck.props.backend.config)});
 
 const serverModule = require('./server');
-const address = main.start(serverModule());
-address.then(function (address) {
+const serverAddress = main.start(serverModule());
+serverAddress.then(function ({ port, address }) {
     if (process && process.send) {
-        process.send(address.port.toString());
+        process.send({ port, address });
     }
 });
-module.exports = address;
+module.exports = serverAddress;
 `;
     }