+++ packages/core/src/browser/widget-open-handler.ts
@@ -15,7 +15,6 @@
  ********************************************************************************/
 
 import { inject, postConstruct, injectable } from 'inversify';
-import { Widget, FocusTracker } from '@phosphor/widgets';
 import URI from '../common/uri';
 import { MaybePromise, Emitter, Event } from '../common';
 import { BaseWidget } from './widgets';
@@ -90,45 +89,11 @@ export abstract class WidgetOpenHandler<W extends BaseWidget> implements OpenHan
         if (!widget.isAttached) {
             this.shell.addWidget(widget, op.widgetOptions || { area: 'main' });
         }
-        const promises: Promise<void>[] = [];
         if (op.mode === 'activate') {
-            promises.push(this.onActive(widget));
-            promises.push(this.onReveal(widget));
-            this.shell.activateWidget(widget.id);
+            await this.shell.activateWidget(widget.id);
         } else if (op.mode === 'reveal') {
-            promises.push(this.onReveal(widget));
-            this.shell.revealWidget(widget.id);
+            await this.shell.revealWidget(widget.id);
         }
-        await Promise.all(promises);
-    }
-    protected onActive(widget: W): Promise<void> {
-        if (this.shell.activeWidget === widget) {
-            return Promise.resolve();
-        }
-        return new Promise(resolve => {
-            const listener = (shell: ApplicationShell, args: FocusTracker.IChangedArgs<Widget>) => {
-                if (args.newValue === widget) {
-                    this.shell.activeChanged.disconnect(listener);
-                    resolve();
-                }
-            };
-            this.shell.activeChanged.connect(listener);
-        });
-    }
-    protected onReveal(widget: W): Promise<void> {
-        if (widget.isVisible) {
-            return new Promise(resolve => window.requestAnimationFrame(() => resolve()));
-        }
-        return new Promise(resolve => {
-            const waitForVisible = () => window.requestAnimationFrame(() => {
-                if (widget.isVisible) {
-                    window.requestAnimationFrame(() => resolve());
-                } else {
-                    waitForVisible();
-                }
-            });
-            waitForVisible();
-        });
     }
 
     /**
@@ -165,5 +130,18 @@ export abstract class WidgetOpenHandler<W extends BaseWidget> implements OpenHan
     }
 
     protected abstract createWidgetOptions(uri: URI, options?: WidgetOpenerOptions): Object;
+
+    async closeAll(): Promise<W[]> {
+        const allClosed: W[] = [];
+        await Promise.all(
+            [this.all.map(async widget => {
+                const closed = await this.shell.closeWidget(widget.id);
+                if (closed) {
+                    allClosed.push(closed as W);
+                }
+            })]
+        );
+        return allClosed;
+    }
 
 }